requirements.txt
curl-cffi>=0.5.7
loguru>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.0
pathlib>=1.0.1
PyExecJS>=1.5.1
pymysql>=1.0.2
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

